WebForCSPM/server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
PyJWT==2.8.0
bcrypt==4.0.1
pymongo==4.5.0
gunicorn==21.2.0
openpyxl==3.1.2
joblib==1.5.1
